requirements.txt
aiofile==1.5.2
aiofiles==0.4.0
aiohttp==3.6.2
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cssselect==1.1.0
dnspython==1.16.0
email-validator==1.0.5
fastapi==0.45.0
fire==0.2.1
graphene==2.1.8
graphql-core==2.2.1
graphql-relay==2.0.1
h11==0.8.1
httptools==0.0.13
idna==2.8
importlib-metadata==1.3.0
inflect==3.0.2
itsdangerous==1.1.0
Jinja2==2.10.3
lxml==4.4.1
MarkupSafe==1.1.1
more-itertools==8.0.2
multidict==4.6.1
parse==1.12.1
promise==2.2.1
pydantic==1.2
pyquery==1.4.1
pysqlcipher3==1.0.3
python-multipart==0.0.5
pywebcopy==6.0.0
PyYAML==5.2
requests==2.22.0
Rx==1.6.1
six==1.13.0
soupsieve==1.9.5
sqlacodegen==2.1.0
SQLAlchemy==1.3.12
starlette==0.12.9
termcolor==1.1.0
ujson==1.35
urllib3==1.25.7
uvicorn==0.10.8
uvloop==0.14.0
w3lib==1.21.0
websockets==8.1
xxtea==1.3.0
yarl==1.3.0
zipp==0.6.0
